+++ packages/adapters-library/src/scripts/adapterBuilder/buildIntegrationTests.ts
@@ -4,7 +4,7 @@ import partition from 'lodash/partition'
 import { parse, print, types, visit } from 'recast'
 import { lowerFirst } from '../../core/utils/caseConversion'
 import { writeAndLintFile } from '../../core/utils/writeAndLintFile'
-import { testCases } from '../templates/testCases'
+import { testCases } from './templates/testCases'
 import { fileExists } from '../utils/fileExists'
 import { sortEntries } from '../utils/sortEntries'
 import n = types.namedTypes
