+++ packages/adapters-library/src/scripts/adapterBuilder/newAdapterCliLogo.ts
