+++ packages/adapters-library/src/scripts/adapterBuilder/newAdapterCommand.ts
@@ -3,14 +3,14 @@ import path from 'node:path'
 import chalk from 'chalk'
 import { Command } from 'commander'
 import { prompt } from 'inquirer'
-import { Chain } from '../core/constants/chains'
-import { lowerFirst, pascalCase } from '../core/utils/caseConversion'
-import { logger } from '../core/utils/logger'
-import { writeAndLintFile } from '../core/utils/writeAndLintFile'
-import type { DefiProvider } from '../defiProvider'
-import { addProtocol } from './adapterBuilder/addProtocol'
-import { buildIntegrationTests } from './adapterBuilder/buildIntegrationTests'
-import { exportAdapter } from './adapterBuilder/exportAdapter'
+import { Chain } from '../../core/constants/chains'
+import { lowerFirst, pascalCase } from '../../core/utils/caseConversion'
+import { logger } from '../../core/utils/logger'
+import { writeAndLintFile } from '../../core/utils/writeAndLintFile'
+import type { DefiProvider } from '../../defiProvider'
+import { addProtocol } from './addProtocol'
+import { buildIntegrationTests } from './buildIntegrationTests'
+import { exportAdapter } from './exportAdapter'
 import { newAdapterCliLogo } from './newAdapterCliLogo'
 import {
   BlankAdapterOutcomeOptions,
