+++ packages/adapters-library/src/scripts/adapterBuilder/replacements.ts
