+++ packages/adapters-library/src/scripts/adapterBuilder/templates/compoundV2BorrowMarketForkAdapter.ts
@@ -1,11 +1,9 @@
-import { Chain } from '../../core/constants/chains'
-import { BlankAdapterOutcomeOptions, QuestionAnswers } from '../questionnaire'
+import { Chain } from '../../../core/constants/chains'
 
-export function uniswapV2PoolForkAdapterTemplate({
+export function compoundV2BorrowMarketForkAdapterTemplate({
   protocolKey,
   adapterClassName,
   productId,
-  chainKeys,
 }: {
   protocolKey: string
   adapterClassName: string
@@ -13,10 +11,7 @@ export function uniswapV2PoolForkAdapterTemplate({
   chainKeys: (keyof typeof Chain)[]
 }) {
   return `
-  import {
-    UniswapV2PoolForkAdapter,
-    UniswapV2PoolForkPositionStrategy,
-  } from '../../../../core/adapters/UniswapV2PoolForkAdapter'
+  import { CompoundV2BorrowMarketForkAdapter } from '../../../../core/adapters/CompoundV2BorrowMarketForkAdapter'
   import { Chain } from '../../../../core/constants/chains'
   import { CacheToFile } from '../../../../core/decorators/cacheToFile'
   import { NotImplementedError } from '../../../../core/errors/errors'
@@ -26,51 +21,27 @@ export function uniswapV2PoolForkAdapterTemplate({
   } from '../../../../types/adapter'
   import { Protocol } from '../../../protocols'
   
-  export class ${adapterClassName} extends UniswapV2PoolForkAdapter {
+  export class ${adapterClassName} extends CompoundV2BorrowMarketForkAdapter {
     productId = '${productId}'
+  
+    contractAddresses: Partial<Record<Chain, { comptrollerAddress: string }>> = {}
   
     getProtocolDetails(): ProtocolDetails {
       return {
         protocolId: this.protocolId,
         name: '${protocolKey}',
-        description: '${protocolKey} pool adapter',
+        description: '${protocolKey} borrow market adapter',
         siteUrl: '',
         iconUrl: '',
-        positionType: PositionType.Supply,
+        positionType: PositionType.Borrow,
         chainId: this.chainId,
         productId: this.productId,
       }
     }
   
-    protected chainMetadataSettings(): Partial<
-      Record<Chain, UniswapV2PoolForkPositionStrategy>
-    > {
-      // TODO - For each supported chain, provide the settings needed to build the list of pools
-      // If using subgraph (recommended for forks with an available subgraph), provide the subgraph URL and factory cotract address
-      // If using factory contract (recommended when subgraph is no available), provide the factory contract address
-      return {
-        ${chainKeys
-          .map((chainKey, i) => {
-            const metadataBuilderType =
-              i === 0
-                ? `type: 'graphql',
-          subgraphUrl:
-            'https://api.thegraph.com/subgraphs/name/<SUBGRAPH-PATH>',
-          factoryAddress: '<FACTORY-CONTRACT-ADDRESS>',`
-                : `type: 'factory',
-              factoryAddress: '<FACTORY-CONTRACT-ADDRESS>',`
-
-            return `[Chain.${chainKey}]: {
-          ${metadataBuilderType}
-        }`
-          })
-          .join(',')}
-      }
-    }
-  
     @CacheToFile({ fileKey: 'protocol-token' })
     async buildMetadata() {
-      return super.buildMetadata()
+      return await super.buildMetadata()
     }
 
     /**
