+++ packages/adapters-library/src/scripts/adapterBuilder/templates/compoundV2SupplyMarketForkAdapter.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../core/constants/chains'
+import { Chain } from '../../../core/constants/chains'
 import { BlankAdapterOutcomeOptions, QuestionAnswers } from '../questionnaire'
 
 export function compoundV2SupplyMarketForkAdapterTemplate({
