+++ packages/adapters-library/src/scripts/adapterBuilder/templates/testCases.ts
