+++ packages/adapters-library/src/scripts/adapterBuilder/templates/writeOnlyDeFiAdapter.ts
@@ -1,4 +1,4 @@
-import { Chain } from '../../core/constants/chains'
+import { Chain } from '../../../core/constants/chains'
 import { BlankAdapterOutcomeOptions, QuestionAnswers } from '../questionnaire'
 
 export function writeOnlyDeFiAdapter({
