+++ packages/adapters-library/src/scripts/index.ts
@@ -7,9 +7,9 @@ import { buildMetadata } from './buildMetadata'
 import { buildMetadataDb } from './buildMetadataDb'
 import { buildSnapshots } from './buildSnapshots'
 import { buildContractTypes } from './buildTypes'
-import { copyAdapter } from './copyAdapter'
+import { copyAdapter } from './adapterBuilder/copyAdapter'
 import { featureCommands } from './featureCommands'
-import { newAdapterCommand } from './newAdapterCommand'
+import { newAdapterCommand } from './adapterBuilder/newAdapterCommand'
 import { performance } from './performance'
 import { simulateTxCommand } from './simulateTxCommand'
 import { stressCommand } from './stress'
